+++ tests/v1/multiprocess/test_mq_handler_helpers.py
@@ -7,7 +7,8 @@ and passed between processes during multiprocessing tests.
 """
 
 # First Party
-from lmcache.v1.multiprocess.custom_types import KVCache, LayoutHints
+from lmcache.v1.gpu_connector.utils import LayoutHints
+from lmcache.v1.multiprocess.custom_types import KVCache
 from lmcache.v1.multiprocess.protocol import KeyType
 
 # ==============================================================================
